+++ IADownloadManager/IASequentialDownloadManager.h
@@ -9,12 +9,12 @@
 #import <Foundation/Foundation.h>
 #import <AFHTTPRequestOperation.h>
 
-typedef void (^IASequentialProgressBlock)(float progress, int index);
-typedef void (^IASequentialCompletionBlock)(BOOL success, AFHTTPRequestOperation *operation, id response, int index);
+typedef void (^IASequentialProgressBlock)(float progress, NSInteger index);
+typedef void (^IASequentialCompletionBlock)(BOOL success, AFHTTPRequestOperation *operation, id response, NSInteger index);
 
 @protocol IASequentialDownloadManagerDelegate <NSObject>
-- (void) sequentialManagerProgress:(float)progress atIndex:(int)index;
-- (void) sequentialManagerDidFinish:(BOOL)success operation:(AFHTTPRequestOperation *)operation response:(id)response atIndex:(int)index;
+- (void) sequentialManagerProgress:(float)progress atIndex:(NSInteger)index;
+- (void) sequentialManagerDidFinish:(BOOL)success operation:(AFHTTPRequestOperation *)operation response:(id)response atIndex:(NSInteger)index;
 @end
 
 @interface IASequentialDownloadManager : NSObject
